feat: Sync on leaving a note
With this change, we send a sync() request to the server
when leaving a note. This will allows the server to always
have files names and content up to date when we finish
to edit a note (instead of having to wait a few minutes).

Collateral: We now go back to the index with an anchor
and not by reloading the full page.