fix: Downgrades Atlaskit + remove codeblocks
HOTFIX
The upgrade to Atlaskit 115 did add `allowCodeBlocks` by default
and removes the possibility to deactivate the plugin.
The stack then fails if someone adds a codeblock in its note
because it doesn't have the corresponding blocks in its schema.

We could upgrade again in the future but this will need to deal
with schema changes.