feat: Public route with share token for a parent folder
This changes add the ability to give an the file id as parameter
in the public route of a share. This is needed when the sharecode
is about multiple files or about a note.

This changes also take into account if the permission of the share code
states the share is read only or not.