fix: Allows catchup from empty documents
Using the jsonTransformer directly introduce a bug for empty documents
with only one paragraph. The transformed document has no content
(not even the paragraph) and makes the Editor to always when think
it is as version=0. This would break the catchup logic.
This patch fixes this unique case, creating a correct empty document with
an empty paragraph