fix: Notes creation on mobile
The update of cozy-client broke the "+" button on mobile.
The AddMobile component uses the withClient / useClient of the
new cozy-client, which relies on the new Context feature of React.
The bar only provides the old Context feature in its tree.
When the AddMobile is inserted into the bar, it cannot find the
cozyClient instance in the tree, and fails.