feat: Listen to Schema update to reload the note
We can be in that situation :
- Alice has opened a note and do not work on it for now but have shared
it with Bob.
- Cozy release a new Note version that change the Schema
- Bob opens the same note, the app is updated, so the schema too,
Bob works on the content.
- Alice goes back to the note but she is still on the old schema and old
app version, so editing can be blocked. (depending on the schema update...)

In order to avoid this issue, when Bob updates the schema, the real
io.cozy.files is updated. So we listen for this EDITED realtime event and
if we detect it, and detect a change in the loaded schema, we reload
the app on Alice's browser.

There is a potential risk of losing the latest editions from Alice (if
Alice network is slow the realtime event can be slow to arrive) but we
accept it since this case should not happen very often.