fix: Do not force `drive` slug when clicking on `BackFromEditing` button
`BackFromEditing` button is responsible to redirect the app to the
`returnUrl` given in the URL's searchParams

In previous behaviour, in a Flagship mobile app, the resulting scenario
would be to call `openApp` with the given URL and the `drive` slug

However, the `drive` slug should not be forced. When the note is opened
from `cozy-notes` then the  `returnUrl` is the `cozy-notes` base url.
So instead we should guess the slug from the `returnUrl`