fix: Font size on iOS devices
We had the note's titles displayed
in a weird shape on iOS devices.

Sometimes the text was small
sometimes it was big... Even if
we have a specified font size
!important.

After a quick search on the internet
it seems that Safari mobile can increase
the font size.

Even if I still don't understand why
we had this behavior, this fix seems to
work. We only target iOS devices and
tell the browser to not set something
else than 100% for ajusting the font size.