fix: adjust table behaviour on text selection
When the user is selecting text inside a cell,
for unknown reasons
AtlasKit will add CSS class `pm-table-last-item-in-cell` along
other DOM elements at unpredictable places in the table.
This specific class will add a margin-bottom
to the last item in a cell to avoid the text to
be too close to the cell border,
but it will also change the height of the cell,
triggering the cursor position of the user to be wrong,
triggering in turn a whole new selection of
text itself triggering a selection of the cell itself, and so on.
Finding the exact origin of the problem is
too time-consuming at this point, so we just
remove the margin-bottom on that class,
this way the table height should not jump
around and the user's cursor should not be
misplaced, avoid the cascade of problems described above.